Data ingestion/requirements.txt
pandas>=2.2.0
requests>=2.32.0
python-dateutil>=2.9.0
tenacity>=9.0.0
beautifulsoup4>=4.12.0
lxml>=5.2.0